+++ CommandLine/WinCronCommandMode.cs
@@ -1,0 +1,10 @@
+namespace WinCron.CommandLine;
+
+public enum WinCronCommandMode
+{
+    Run,
+    Test,
+    List,
+    Version,
+    Help
+}